move PTRenderTest to render test namespace
